Update patch set 1
Patch Set 1:

ok you might not be able to replicate the issue, it doesn't mean there isn't a bug
I would gladly send logs, however I'm unable to personally use cm10 on the l5 at the moment.
I pushed this on behalf of the people who are experiencing issues (both the l5 and l7)
the issue has been traced back to the kernel so far, and I am in the process of checking the code.
However I still feel that even a nightly should not have bugs so serious, as do many others

Patch-set: 1